feat: rename TemporalPaginationService to TemporalPaginationUtils
